old-tests/submission/W3C/domhtml/table48.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLECOL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-tr-align" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<table summary="Table Summary">
<colgroup align="center" char="$" charoff="15" span="2" valign="middle" width="20">
<col align="center" char="*" charoff="20" span="1" valign="middle" width="20">
</colgroup>
<tbody><tr>
<th>Id</th>
<th>Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
<tr>
<td>EMP0001</td>
<td>Martin</td>
<td>Accountant</td>
<td>56,000</td>
</tr>
</tbody></table>











<script>
var domtest = async_test("The align IDL attribute of the tr element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valign;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("col");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      valign = testNode.align;

      assert_equals(valign,"center");
       
    domtest.done();
}));

</script>
<p>The align IDL attribute of the tr element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
